requirements.txt
mysql-connector-python==8.0.32
pandas==2.0.0
numpy==1.24.3
plotly==5.14.1
openpyxl==3.1.2
google-analytics-data==0.16.2 